+++ src/dkg.ts
@@ -19,7 +19,7 @@ import { ChecksumAddress } from './types';
 import { bytesEquals, fromHexString, objectEquals } from './utils';
 
 export function getVariantClass(
-  variant: FerveoVariant
+  variant: FerveoVariant,
 ): typeof DecryptionShareSimple | typeof DecryptionSharePrecomputed {
   if (variant.equals(FerveoVariant.simple)) {
     return DecryptionShareSimple;
@@ -31,9 +31,9 @@ export function getVariantClass(
 }
 
 export function getCombineDecryptionSharesFunction(
-  variant: FerveoVariant
+  variant: FerveoVariant,
 ): (
-  shares: DecryptionShareSimple[] | DecryptionSharePrecomputed[]
+  shares: DecryptionShareSimple[] | DecryptionSharePrecomputed[],
 ) => SharedSecret {
   if (variant.equals(FerveoVariant.simple)) {
     return combineDecryptionSharesSimple;
@@ -61,8 +61,9 @@ export class DkgRitual {
     public readonly id: number,
     public readonly dkgPublicKey: DkgPublicKey,
     public readonly dkgParams: DkgRitualParameters,
-    public readonly state: DkgRitualState
-  ) {}
+    public readonly state: DkgRitualState,
+  ) {
+  }
 
   public toObj(): DkgRitualJSON {
     return {
@@ -74,16 +75,16 @@ export class DkgRitual {
   }
 
   public static fromObj({
-    id,
-    dkgPublicKey,
-    dkgParams,
-    state,
-  }: DkgRitualJSON): DkgRitual {
+                          id,
+                          dkgPublicKey,
+                          dkgParams,
+                          state,
+                        }: DkgRitualJSON): DkgRitual {
     return new DkgRitual(
       id,
       DkgPublicKey.fromBytes(dkgPublicKey),
       dkgParams,
-      state
+      state,
     );
   }
 
@@ -107,25 +108,25 @@ export class DkgClient {
   public static async initializeRitual(
     web3Provider: ethers.providers.Web3Provider,
     ursulas: ChecksumAddress[],
-    waitUntilEnd = false
+    waitUntilEnd = false,
   ): Promise<number | undefined> {
     const ritualId = await DkgCoordinatorAgent.initializeRitual(
       web3Provider,
-      ursulas.sort()
+      ursulas.sort(),
     );
 
     if (waitUntilEnd) {
       const isSuccessful = await DkgClient.waitUntilRitualEnd(
         web3Provider,
-        ritualId
+        ritualId,
       );
       if (!isSuccessful) {
         const ritualState = await DkgCoordinatorAgent.getRitualState(
           web3Provider,
-          ritualId
+          ritualId,
         );
         throw new Error(
-          `Ritual initialization failed. Ritual id ${ritualId} is in state ${ritualState}`
+          `Ritual initialization failed. Ritual id ${ritualId} is in state ${ritualState}`,
         );
       }
     }
@@ -135,7 +136,7 @@ export class DkgClient {
 
   private static waitUntilRitualEnd = async (
     web3Provider: ethers.providers.Web3Provider,
-    ritualId: number
+    ritualId: number,
   ): Promise<boolean> => {
     return new Promise((resolve, reject) => {
       const callback = (successful: boolean) => {
@@ -151,11 +152,11 @@ export class DkgClient {
 
   public static async getExistingRitual(
     web3Provider: ethers.providers.Web3Provider,
-    ritualId: number
+    ritualId: number,
   ): Promise<DkgRitual> {
     const ritualState = await DkgCoordinatorAgent.getRitualState(
       web3Provider,
-      ritualId
+      ritualId,
     );
     const ritual = await DkgCoordinatorAgent.getRitual(web3Provider, ritualId);
     const dkgPkBytes = new Uint8Array([
@@ -169,18 +170,19 @@ export class DkgClient {
         sharesNum: ritual.dkgSize,
         threshold: assumedThreshold(ritual.dkgSize),
       },
-      ritualState
+      ritualState,
     );
   }
 
+
   public static async verifyRitual(
     web3Provider: ethers.providers.Web3Provider,
-    ritualId: number
+    ritualId: number,
   ): Promise<boolean> {
     const ritual = await DkgCoordinatorAgent.getRitual(web3Provider, ritualId);
     const participants = await DkgCoordinatorAgent.getParticipants(
       web3Provider,
-      ritualId
+      ritualId,
     );
 
     const validatorMessages = participants.map((p) => {
@@ -202,13 +204,13 @@ export class DkgClient {
     const participantPublicKeys: Record<string, FerveoPublicKey> = {
       '0x210eeAC07542F815ebB6FD6689637D8cA2689392': FerveoPublicKey.fromBytes(
         fromHexString(
-          'ace9d7567b26dafc512b2303cfdaa872850c62b100078ddeaabf8408c7308b3a43dfeb88375c21ef63230fb4008ce7e908764463c6765e556f9b03009eb1757d179eaa26bf875332807cc070d62a385ed2e66e09f4f4766451da12779a09036e'
-        )
+          'ace9d7567b26dafc512b2303cfdaa872850c62b100078ddeaabf8408c7308b3a43dfeb88375c21ef63230fb4008ce7e908764463c6765e556f9b03009eb1757d179eaa26bf875332807cc070d62a385ed2e66e09f4f4766451da12779a09036e',
+        ),
       ),
       '0xb15d5A4e2be34f4bE154A1b08a94Ab920FfD8A41': FerveoPublicKey.fromBytes(
         fromHexString(
-          '8b373fdb6b43e9dca028bd603c2bf90f0e008ec83ff217a8d7bc006b585570e6ab1ce761bad0e21c1aed1363286145f61134ed0ab53f4ebaa05036396c57f6e587f33d49667c1003cd03b71ad651b09dd4791bc631eaef93f1b313bbee7bd63a'
-        )
+          '8b373fdb6b43e9dca028bd603c2bf90f0e008ec83ff217a8d7bc006b585570e6ab1ce761bad0e21c1aed1363286145f61134ed0ab53f4ebaa05036396c57f6e587f33d49667c1003cd03b71ad651b09dd4791bc631eaef93f1b313bbee7bd63a',
+        ),
       ),
     };
